+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## 0.4.4 (unreleased)
+## 0.5.0 (unreleased)
 
 - Added experimental support for `pg_stat_monitor`
 - Improved output when HypoPG not installed
+- Dropped support for Ruby < 2.7
 
 ## 0.4.3 (2023-03-26)